feat: generate Account Code

+++ src/main/java/kr/ibct/springboilerplate/account/AccountController.java
@@ -2,8 +2,8 @@ package kr.ibct.springboilerplate.account;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -12,6 +12,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.validation.Valid;
 import java.net.URI;
+import java.util.List;
 
 import static org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder.on;
 
@@ -24,8 +25,9 @@ public class AccountController {
 
 
     @PostMapping
-    public ResponseEntity<?> signUp(@RequestBody @Valid AccountDto.signUpRequest signUpRequest, BindingResult errors) {
+    public ResponseEntity<?> signUp(@Valid @RequestBody AccountDto.signUpRequest signUpRequest, Errors errors) {
         if (errors.hasErrors()) {
+
             return ResponseEntity.badRequest().body(errors);
         }
 
@@ -38,27 +40,65 @@ public class AccountController {
         URI uri = uriComponents.encode().toUri();
         return ResponseEntity.created(uri).body(signUpRequest);
     }
+
+    @PostMapping("/token")
+    public ResponseEntity<?> generateToken(@RequestBody AccountDto.SignInRequest request, BindingResult errors) {
+        if (errors.hasErrors()) {
+            return ResponseEntity.badRequest().body(errors);
+        }
+        String token = accountService.provideToken(request.getEmail(), request.getPassword());
+        return ResponseEntity.ok(token);
+    }
+
+    @GetMapping
+    public ResponseEntity<?> getAllAccount() {
+        List<Account> allAccount = accountService.getAllAccount();
+        return ResponseEntity.ok(allAccount);
+    }
+
     @GetMapping("/{id}")
-    public ResponseEntity<?> getAccount(@PathVariable Long id) {
-        AccountDto.accountResponse account = accountService.getAccount(id);
-        return ResponseEntity.ok(account);
+    public ResponseEntity<?> getAccount(@PathVariable Long id, @CurrentUser Account account) {
+        if (!verifyAccount(id, account)) {
+            return ResponseEntity.badRequest().body("bad Request");
+        }
+        return ResponseEntity.ok(accountService.getAccount(id));
     }
 
     @PatchMapping("/{id}")
     public ResponseEntity<?> updateAccount(@PathVariable Long id,
                                            @Valid @RequestBody AccountDto.updateRequest request,
-                                           BindingResult errors){
+                                           @CurrentUser Account account,
+                                           BindingResult errors) {
         if (errors.hasErrors()) {
             return ResponseEntity.badRequest().body(errors);
         }
-        accountService.updateAccount(id,request);
+        if (!verifyAccount(id, account)) {
+            return ResponseEntity.badRequest().body("bad Request");
+        }
+        accountService.updateAccount(id, request);
         return ResponseEntity.ok().body("");
     }
 
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteAccount(@PathVariable Long id) {
+    public ResponseEntity<?> deleteAccount(@PathVariable Long id, @CurrentUser Account account) {
+        if (!verifyAccount(id, account)) {
+            return ResponseEntity.badRequest().body("bad Request");
+        }
         accountService.deleteAccount(id);
         return ResponseEntity.noContent().build();
+    }
+
+
+    // Todo AOP or interceptor로 변경하기
+    private boolean verifyAccount(Long id, Account account) {
+        if (account.getRoles().contains(AccountRole.ADMIN)) {
+            return true;
+        }
+        if (account.getId() == id) {
+            return true;
+        }
+        return false;
     }
